feat: remove HomePage implementation to prepare for new landing screen
- Add MyBookList component to display logged-in user's book list
